+++ changuitasAppFront/auxiliares/utils.ts
@@ -1,0 +1,3 @@
+export const capitalizarPrimeraLetra = (texto: string): string => {
+  return texto.charAt(0).toUpperCase() + texto.slice(1);
+};
